book: initial CSS structure
This will be a pain, as we'll need to replicate CSS stylesheet changes
across all repos/books, rather than fetching from a common source on the
web. The reason for this approach is that the whole release package
should be whole, and consumable without an internet connection.

In future work, print media queries will allow us to tune the print form
of these books, and can be used to generate PDF versions.