Regenerate images in Book 1 chapter 9 as PNGs instead of JPGs with artifacts (#179).
I ran PNGs through pngquant+Zopfli to produce smallest possible sizes, but each image is several kilobytes larger than the previous JPG.

The images are produced with the diffuse fix (#181).

Also fixed the example in chapter 9 to compile (metal does not have fuzz parameter yet), and use a similar red color as in the image, instead of blue.