color: large update to color handling
This change rolls back the prior commit that created a new `color`
class. I ran across this section in the text, which explains the
original rationale for using `vec3` for everything:

    We’ll use the same class `vec3` for colors, locations, directions,
    offsets, whatever. Some people don’t like this because it doesn’t
    prevent you from doing something silly, like adding a color to a
    location. They have a good point, but we’re going to always take the
    "less code" route when not obviously wrong.

Seems good enough reason for me, so I removed the short-lived `color`
class and instead added the `vec3::write_color()` method. In the end,
this is quite a bit less code, and fullfills everything we needed in one
simple function. (Ok, two, if you count the new `clamp()` utility.)

This change includes source and all text changes.

In addition, this change sneaks in a slight change in the output image
parameters. On my setup, a release build now takes about 2-3 minutes for
each of the three programs, which makes sanity perf testing easier.