Clarify sphere intersect + slight perf improvement
This change updates both the books and the code.

The sphere intersection code used the fact that the b (in the quadratic
formula) had a factor of two in it. This yielded a simplification in the
formula that hinted at, but not explained. Further, the variable in the
new formula used the unfortunate name `b`, when it was really b/2.

This change adds a deriviation of the simplification, and changes the
code to use a more honest name, and also adds some very tiny efficiency
improvements.

Resolves #113