hitvec restLife: hittable_list uses vector
For bvh.h, the original code for TheRestOfYourLife had changes in it
plus a final commented-out section. Once that's removed, most of the new
code is pointless. In addition, the book makes no mention of BVH
changes. In the end, I've decided to make the code identical to match
the text.