Improve Perlin class initialization
The prior form defined a private `init()` method to initialize the
Perlin class. This had several drawbacks. First, there really wasn't a
need for the indirection of calling a separate method from the
constructor. Second, because the `init()` method did not delete any
existing allocations, it was not safe to call more than once.

Given all this, there was really no good reason to not just perform the
initialization in the constructor.