Move acknowledgments to single common document
The prior version had an identical acknowledgments section copy and
pasted to the end of each book, a solution that was both tedious and
error prone.

This change uses Markdeep's include directive to import the section
from a new single `acknowledgments.md.html` file.

Additionally, this change uses Markdeep's auto reference, so that
`[acknowledgments](#acknowledgements)` is now replaced with the text
"acknowledgments section", which Markdeep auto links.