Update camera code
In the prior code, changes to camera variables would require associated
changes to a number of other related variables, in ways that were not
apparent. This change uses variables a more consistent manner, where
changing any particular variable should yield consistent results in
other values.

In addition, this provides a more consistent series of iterative steps
as the camera develops through the text.

Resolves #527