Fix `random_unit_vector()`
I added an alternate formula for uniformly-distributed points on a unit
sphere for this function. It's presented in
https://mathworld.wolfram.com/SpherePointPicking.html, equations 6-8.

Unfortunately, I misinterpreted "u in [-1, 1]" to be equivalent to
`u = random_double(-1,1)`. However, you're supposed to pick u = cos(phi)
to be uniformly distributed.

I replaced this body with the original one, just generating a unit point
_in_ the unit sphere, and normalizing the result.

Resolves #697