Refactor sphere hit method to use common code
The sphere::hit() function calls get_sphere_uv() with the intersection
point transformed to its position on a unit sphere centered at the
origin. But this calculation was already performed for the normal
vector, so we should just reuse that result instead of calculating it
twice.

Also, the prior sphere::hit() methods duplicated the main update block
with a single change in variable value. This change solves and selects
the intersection root first, then updates the hit record for a good hit,
simplifying the code and the text.