Untested: ray_color takes in lights
The prior code had multiple implementations of `ray_color()` that
created a new light on every ray bounce. This change now passes in the
light shape from the scene.

NOTE: This change is untested. We have big changes pending for book
three when we begin v4.0.0 development soon, so I expect that this code
will be heavily modified (and tested) at that time.

Resolves #759