source: Implement new scene class
This is a major change to the codebase, refactoring multiple components
that were previously uncoordinated.

Generally, we introduce a new `scene` class, which is responsible for
managing image parameters, the camera, the world geometry, and lights.
This is particularly useful in the book 2 code, where we maintain a
number of different scenes, with the ability to choose & render any one
of them. Since all of the above properties can be unique to each scene,
it became cumbersome to ensure that all parts were properly set up. The
new `scene` class makes this significantly easier, and also simplifies
the render code as well.

The camera class is changed so that camera parameters are exposed as
public member variables, and may be updated in random order as needed. A
new `camera::initialize()` method now explicitly sets the camera up for
rendering after parameters are set, and before `camera::get_ray` is used
to generate scene rays.

Aspect ratio is maintained now by the `scene` class, so that it
coordinates the rendered image dimensions and the camera viewport.

The `ray_color()` function is now located in the `scene` class as a
private method, and the render loop that used to be in the `main()`
functions is now moved into the `scene` class as a public method. This
also allows some coordination between the two methods at render time.

An additional change tucked into this commit is the fact that the book 2
code changes the default scene from the Cornell box to the book 2 final
scene with simplified parameters for a faster render. This scene
exercises significantly more of the book 2 code, so also now serves as a
good sanity and performance test.