rtw_stb_image load searches RTW_IMAGES _and_ rest
Changed the image file search strategy to search _both_ the directory
specified with the RTW_IMAGES environment variable _and_ the other
possible paths.

In addition, I've removed the STBI error reporting, since with any
format or other load issue, we'll just fail and then look in the next
speculative directory. The result is that the _last_ error is reported,
which is most likely that the file just didn't exist at that location.

I couldn't find a way to get more programmatic error information in
order to stop at the first non-existence load failure. I would like to
have used the `std::filesystem` library to stop once I found the first
file that existed (and _then_ try the load), but `std::filesystem` is
C++17, and I'd rather not move from C++11 until we really need to.