aabb ctor treats 2 params as extrema in any order
The original constructor expected that the first parameter was the
minimum point, and the second was the maximum. The new constructor now
selects the minimums of all coordinates in the two parameters, and the
maximums of all the coordinates. In this way, the two parameters are
just two extreme points of the bounds, so is significantly more robust.

This change was done because the original code had bugs when handling
spheres of negative radius (used to model glass spheres with voids).

Resolves #733