scene: parameterize x resolution and samples/pixel
These two parameters are better managed by the client of the `scene`
class, as they are purely about the tradeoff between time and quality.

I'm leaving max_depth to the scene, as different scenes reasonably have
different maximum ray depth requirements. For example, a purely diffuse
scene in a Cornell box doesn't require as many bounces as a scene with
lots of reflection and refraction.

In addition, I'm changing from image_width to resolution_x, as width
could mean either the width of the image viewport in 3D space, or it
could mean the integral count of pixels horizontally. This will also
help keep concepts clean in the camera class.

I'm leaving updates to the books and the other two sources for following
commits to make reviewing easier.