aabb: fix performance puzzle with aabb::hit()
It turns out that the slow version, preserved in the original code, had
a bug in that it failed to tighten the `ray_t` interval with each
successive iteration through the X, Y and Z axes. Instead, it was
modifying temporary variables.

This change cleans up the two versions of the `hit()` method, and aligns
the text's version with our codebase.

Resolves #817