Fix sphere intersection test
This is in response to issues raised in issue #875, where the sphere
rendering image appears darker than it used to. The issue is that points
that lie exactly on the ray interval used to be excluded, but then were
not. The issue was compounded when we started to use the
`interval::contains()` method, because this subtlety was one step
removed.

This change adds a new `interval.surrounds()` method that tests for
points on the open (instead of closed) interval.

This should fix the issue as raised, but we will verify in the v4.0.0
release milestone, when we regenerate all of the images.

See issue #875
See issue #1012