Tweak sphere::hit() method to address dark shading
This is in response to issues raised in issue #875, where the sphere
rendering image appears darker than it used to. The issue is that points
that lie exactly on the ray interval used to be excluded, but then were
not. The issue was compounded when we started to use the
`interval::contains()` method, because this subtlety was one step
removed.

This should fix the issue as raised, but we will verify in the v4.0.0
release milestone, when we regenerate all of the images.