Use hittable instead of hittable_list if possible
The camera class was using `hittable_list` for world and lights, but
didn't really need to know whether it was a single or multiple
hittables (it only used the `hit()` method). Thus, this changes the
camera methods to use the more general reference.

Similarly, `hittable_pdf` took a reference to a `hittable_list`, but it
only needed a reference to the more general `hittable` class.