Changelog (#1194)
* Update CHANGELOG to 2023-07-28

This is still in rough form, organized only by order of change. A
rewrite of the entire v4 section is still needed for final version for
the upcoming v4.0.0-alpha.1 release.

* Groom alpha.1 section of CHANGELOG + general clean

This adds the release blurb for the alpha.1 changelog entry, and grooms
the pile of line-items into final form.

In addition, I've reworked the entire changelog to present line items in
the following order: fixes, changes, new, and deleted.