Revise book 1 next steps
I've updated the "Next Steps" section at the end of book 1. It included
a mix of topics addressed in book 2 along with topics we haven't yet
covered in any way.

I've split this into topics the reader can find in book 2, a description
of the content of book 3, and a list of uncovered topics.

Resolves #1209