Update reference to Foley, van Dam
The book currently refers to "Fundamentals of Interactive Computer
Graphics". That was certainly one of the seminal books of the field but
it has been rewritten by the same (and some more) authors since it was
first published in 1982 and renamed "Computer Graphics: Principles and
Practice". I think it is a lot easier to find copies of the newer
editions today, they cover more topics, are more up-to-date in general.

I think it makes sense to refer to the book by it's newer name (well, it
was new in 1990). There is a brief history of the book at
https://en.wikipedia.org/wiki/Computer_Graphics:_Principles_and_Practice.