Reverse ray-sphere direction and calculations
Trevor observed that the equation for the ray-sphere intersection was
backwards in that it used (P - C) [center to ray origin] instead of
(C - P) [ray origin to sphere center].

That's all well and good, but it ends up rippling all the way back to
the fundamental sphere equations presented in the text.

Sigh. And I couldn't resist.

In flipping everything so it flows without an odd inversion in the
middle, I also discovered some annoying minus signs that could be
eliminated, plus some bits that needed a tiny bit more explanation.
Quite a bit more work than it originally appeared, but I'm pleased with
the results.

Resolves #1191