perlin::turb() now requires both parameters
The original version defaulted the `depth` parameter to 7.

This was the second instance of default parameters in the book source.
The first was for the time value in the `ray` constructor, which we
addressed by implementing two constructors.

Generally, we're leaning away from using default parameter values in the
interest of simplicity and clarity.