Simplify vec3 initialization
We were using brace initialization like so:

    vec3() : e{0,0,0} {}
    vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}

Some readers reported issues on Mac using braced initialization, so
likely an issue with some compiler version. It may be technically legal
with C++11 (I think, or perhaps a later C++ version), but it's not
really buying anything, so this change uses an older-style
initialization.

Resolves #847