Remove pixel sampling knowledge from write_color
The prior version was passed the number of samples for the summed pixel
color, which it then had to divide out in order to derive the average
sample color.

This change simplifies the write_color function to take only a color,
requiring the caller to compute any desired final color.

In the service of this, the `camera` class now caches the inverse of the
number of pixel samples, so it can compute the average sample color
before passing the result to the now simplified write_color() function.

In addition, I've factored out some steps in the final color value
computation to make each phase more discrete and easier to grok.