Correct call to reflectance()
The function `reflectance()` takes `double ref_idx` (which is `ir` in
the outer scope), not `refraction_ratio`.

Confirmed with the formula for Schlick's approximation
(https://en.wikipedia.org/wiki/Schlick%27s_approximation) which uses two
indexes of refraction, of which one is hard coded to `1` in the function
here.

A future todo would be to keep track of the "ir of the current medium"
in the ray and use that instead of a hard coded `1` in the reflectance
formula.

Originally due to https://github.com/TheThief.