Delete reflection vector normalization call
When computing the reflected ray, we had code to normalize the result.
This is unnecessary, as the reflected ray result will have the same
length as the incoming ray, which should be fine.

In addition, I've added some clarification around the computation of the
magnitude of the vector b used in the reflection computation.

Resolves #1340