Normalize constructor parameter names
In order to avoid confusion between constructor parameter names and
member variables, we either used alternate names for the parameters or
prefixed the parameter names with an underscore. For example:

    class foo {
      public:
        foo(int _i, bool semaphore, short bar) : i(_i), flag(semaphore)
        {
            baz = bar;
        }
       private:
         int i;
         bool flag;
         short baz;
    };

However, this is unnecessary. Constructor initializer lists are
unambiguous, as the assigned item always refers to the class member, and
the assigned value is always from the parameter (unless prefixed with
"this->").

Inside the constructor body, parameter names override member variables,
and "this->" can be used to indicate the class member in the case of
name collision.

Thus, the above code can be rewritten as this:

    class foo {
      public:
        foo(int i, bool flag, short baz) : i(i), flag(flag)
        {
            this->baz = baz;
        }
       private:
         int i;
         bool flag;
         short baz;
    };

I've taken advantage of this to clarify names used in constructors
throughout the codebase. Mostly for constructor parameter names,
sometimes for member variable names.