Implement bubbles with IOR, not negative radii
We were using negative radii on spheres to invert the sense of front vs
back faces of spheres, and using this to get an inverted index of
refraction (IOR) for bubbles. This method was used to model a hollow
glass sphere.

Over the past couple of years we've encountered a number of bugs induced
by negative radii spheres (see the list in the description for issue
1420).

Instead, this change interprets the dielectric parameter not as an IOR
relative to surrounding air, but instead as the ratio of the IOR of the
object material over the IOR of the surrounding medium. Interpreting the
parameter this way lets you model a sphere of air enclosed in glass --
just what we need to model a hollow glass sphere purely with the proper
IORs, and without resorting to inverted geometry with negative radii
spheres.

We also end up using this method to properly demonstrate total
internal/external reflection in another section.

Resolves #1420