Fix section describing total internal reflection
With a sphere of index of refraction greater than the surrounding
medium, there is no incoming ray that can be chosen to show total
internal reflection. An incoming ray will be bent to a smaller angle
relative to the surface normal, and then bent back to the original angle
on exit.

To illustrate total reflection, this change renders a bubble of air
inside water. Glancing rays are then reflected off the surface instead
of refracted.

In the service of this change, I've also moved the center sphere back
a bit from the left and right spheres so you can see the affected side
more clearly. This will invalidate all of the other three-sphere images,
which will be fixed up in #1358 (Alpha.2 milestone verify Book 1
progression).

Resolves #900