Load images in linear gamma
We were using the STBI `stbi_load()` function, which loads
gamma-corrected byte data from the given image.

Instead, we want our image data to be in linear space for internal
computation, with a final gamma=2 correction applied when we write the
rendered result.

Happily, there's an `stbi_loadf()` function that loads linearly-encoded
image data as triples of 32-bit floating-point values.

We don't really need to explain all the machinery, as we kind of just
plop `rtw_stb_image.h` down for the readers to use. I do add some
explanation though that we want loaded images to be in linear color
space.

Resolves #842