Workaround for Markdeep bug in code listings
It turns out that a pattern like '<a b>' inside a code listing fools
Markdeep into thinking it has an embedded HTML tag or something, and
such a string will be expanded to '<a b="">'. Thus the line

    *bptr = static_cast<unsigned char>(*fptr * 256.0);

gets transformed into

    *bptr = static_cast<unsigned char="">(*fptr * 256.0);

when rendered.

I've sent email to Morgan about this bug, but in the meantime, the best
workaround I could come up with that can be safely copy-pasted and
compiled is

    *bptr = static_cast< unsigned char >(*fptr * 256.0);

That is, inserting spaces in the angle brackets. Somewhat ugly, but
thankfully this is in rtw_stb_image.h, which the reader just uses as a
gift instead of really working with.

Issue #1463 tracks the work to undo this once it's fixed in Markdeep.