Address HTML parsing bug with '<unsigned char>'
When listing text contains something like

  static_cast<unsigned char>(thing)

HTML parses `<unsigned char>` as an HTML tag before Markdeep gets a
chance to see it. We had temporarily addressed this by adding spaces
inside the angle brackets, and sent an inquiry off to Morgan McGuire,
creator of Markdeep. His response was that Markdeep couldn't address
this given the parsing order, and to just wrap the entire listing in

  <script type="preformatted">
    ...
  </script>

Resolves #1463