Use explicit std:: namespace almost everywhere
When referring to C++ standard library functions, use the explicit
`std::` namespace qualifier. The only exception is that we bring
`shared_ptr` and `make_shared` into the global namespace in order to
keep the code brief, since we have a 96-character line limit in our
source code.

Currently, this covers the following identifiers:

    acos        fixed          numeric_limits    sqrt
    atan2       floor          ostream           string
    cerr        flush          rand              tan
    clog        fmax           setprecision      vector
    cos         fmin           shared_ptr
    cout        log            sin
    fabs        make_shared    sort

Resolves #1487