Remove unused operator<< & const-medium debug code
When I created issue #1495, I must have been in the middle of a code
progression, because it appeared to me that vec3::operator<< was unused.

Turns out it _is_ used, in debug code for the `constant_medium::hit()`
function. This code was created as a clean-up for the original scattered
debug statements in this function. On review, however, this appears a
bit odd, as the text does not mention the debuggability option (it's
off by default), and no other code in the books has debug print code.

Looking at it again, I think it should be removed, and readers are free
to add whatever debug code they need at any stage in development.

Resolves #1495