Simplify the ONB class
The `onb` class implemented several unused methods, so I've removed them
in the interest of simplicity. These include the operator[] methods to
retrieve the basis axes, as well as a transform function that took three
individual double values. Finally, instead of having an uninitialized
form with a `build_from_w()` function, the class now requires a vector
at construction time, consistent with our use of the class.

I've renamed `local()` to `transform()` for a bit more clarity, without
trying to name either the input space or the result space, as "local"
means different things at different times, and to different people.

Resolves #1088